+++ CHANGELOG.md
@@ -16,7 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Cylinder: random_point_inside, interference_volume_with_other_cylinder, lhs_points_inside
 * CylindricalSurface3D: line_intersections, linesegment_intersections, plane_intersection
 * Line2D: point_distance
+
 * Line3D: to_2d
+* ArcEllipse2D: straight_line_area
+* Line3D: skew_to: Verifies if two Line3D are skew
+* LineSegment3D.line_interserctions
 
 
 
@@ -27,10 +31,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fix some to_step methods from edges.py and faces.py
 * Cylinder: point_belongs
 * FullArc3D: plot (use discretization_points instead of discretise)
+* Face3D.line_intersections: consider borders
 * STL: from stream (use BinaryFile and StringFile instead of io.BinaryIO and FileIO)
 * Step: from stream (use BinaryFile instead of io.BinaryIO)
 * Contour: is_overlapping (consider intersecting_points is empty)
 * LineSegment2D: to_wire (use discretization_points instead of discretise)
+
 
 
 ### Performance improvements
@@ -38,7 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Avoid unneeded bbox computation
 
 ### Refactorings
-
+* LineSegment3D.intersections
 
 
 ### Unittests
@@ -50,6 +56,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Cylinder: point_belongs, random_point_inside, interference_volume_with_other_cylinder, min_distance_to_other_cylinder, is_intersecting_other_cylinder, lhs_points_inside
 * CylindricalFace3D: linesegment_intersections
 * CylindricalSurface3D: line_intersections
+* Line3D: line_distance
+* Line3D: skew_to
+* Line3D: intersections
+* LineSegment3D: line_intersections
+* LineSegment3D: linesegment_intersections
 * Contour: is_overlapping
 * LineSegment2D: to_wire